requirements.txt
fastapi
uvicorn
numpy
joblib
rapidfuzz
rdkit-pypi
pubchempy
